initial dynamic setup for dynamic fields added

+++ client/src/pages/users/admin/pages/AdminCreateForm/FormUtils/elements/DraggableItem.jsx
@@ -2,33 +2,38 @@ import React from 'react';
 import { useDrag } from 'react-dnd';
 import './DraggableItem.css';
 
-const DraggableItem = ({ id, text, dropped, onDelete }) => {
+const DraggableItem = ({ id, text, dropped, onDelete, component: Component }) => {
   const [{ isDragging }, dragRef] = useDrag(() => ({
     type: 'item',
-    item: { id, text },
+    item: { id, text, component: Component },
     collect: (monitor) => ({
       isDragging: monitor.isDragging(),
     }),
   }));
 
-  return (
-    <div
-      ref={dragRef}
-      className="draggable-item"
-      style={{ opacity: isDragging ? 0.5 : 1 }}
-    >
-      <span className='dragItemTitle'>{text}</span>
-      {dropped && (
+  if (dropped) {
+    return (
+      <div className="dropped-item">
+        <input type="text" className='heading-text' placeholder={text} />
         <div className='dragged-item'>
-        {/* <input type="button" value="submit" /> */}
-        <button   onClick={() => onDelete(id)} className="delete-button">
-          ❌
-        </button>
+          {Component && <Component />}
+          <button onClick={() => onDelete(id)} className="delete-button">
+            ❌
+          </button>
         </div>
- 
-      )}
-    </div>
-  );
+      </div>
+    );
+  } else {
+    return (
+      <div
+        ref={dragRef}
+        className="draggable-item"
+        style={{ opacity: isDragging ? 0.5 : 1 }}
+      >
+        <span className='dragItemTitle'>{text}</span>
+      </div>
+    );
+  }
 };
 
 export default DraggableItem;
